requirements.txt
# Market Analyzer Pro v2 — Dependencies

# ── Web Framework ───────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Database ────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0

# ── Task Queue ──────────────────────────────────────
celery[redis]>=5.3.0

# ── Data & Analysis ─────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
pandas-ta>=0.3.14b1
vectorbt>=0.26.0
scipy>=1.12.0

# ── Data Sources ────────────────────────────────────
yfinance>=0.2.36
ccxt>=4.2.0
httpx>=0.26.0
requests>=2.31.0
praw>=7.7.0
websockets>=12.0

# ── NLP / Sentiment ─────────────────────────────────
# FinBERT runs as microservice — deps are in services/finbert/requirements.txt
# TextBlob kept as fallback:
textblob>=0.18.0

# ── Monitoring ──────────────────────────────────────
prometheus-fastapi-instrumentator>=6.1.0
structlog>=24.1.0

# ── Utilities ───────────────────────────────────────
python-dotenv>=1.0.0
python-dateutil>=2.8.2
orjson>=3.9.10
tenacity>=8.2.0
python-telegram-bot>=20.7

# ── Async ───────────────────────────────────────────
aiofiles>=23.2.0
